remove cyclic dependency
