+++ ethereum/eip712/encoding.go
@@ -25,7 +25,7 @@ import (
 	txTypes "github.com/cosmos/cosmos-sdk/types/tx"
 
 	apitypes "github.com/ethereum/go-ethereum/signer/core/apitypes"
-	"github.com/evmos/ethermint/app"
+	appparams "github.com/evmos/ethermint/app/params"
 	"github.com/evmos/ethermint/types"
 
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -40,7 +40,7 @@ var (
 // The process of unmarshaling SignDoc bytes into a SignDoc object requires having a codec
 // populated with all relevant message types. As a result, we must call this method on app
 // initialization with the app's encoding config.
-func SetEncodingConfig(cfg app.EncodingConfig) {
+func SetEncodingConfig(cfg appparams.EncodingConfig) {
 	aminoCodec = cfg.Amino
 	protoCodec = codec.NewProtoCodec(cfg.InterfaceRegistry)
 }
